+++ src/commands/hansolo-cleanup.ts
@@ -2,61 +2,213 @@ import { CommandHandler } from './types';
 import { SessionRepository } from '../services/session-repository';
 import { GitOperations } from '../services/git-operations';
 import { ConsoleOutput } from '../ui/console-output';
-import { ProgressIndicator } from '../ui/progress-indicators';
-import { BoxFormatter } from '../ui/box-formatter';
+import { WorkflowProgress } from '../ui/progress-indicators';
+import { PreFlightChecks, PostFlightChecks, CheckResult } from '../services/validation/pre-flight-checks';
+import { getBanner } from '../ui/banners';
 import chalk from 'chalk';
 
-export class HansoloCleanupCommand implements CommandHandler {
+/**
+ * Pre-flight checks for cleanup command
+ */
+class CleanupPreFlightChecks extends PreFlightChecks {
+  constructor(
+    private gitOps: GitOperations,
+    private sessionRepo: SessionRepository
+  ) {
+    super();
+    this.setupChecks();
+  }
+
+  private setupChecks(): void {
+    this.addCheck(async () => this.checkGitRepository());
+    this.addCheck(async () => this.checkHansoloInitialized());
+  }
+
+  private async checkGitRepository(): Promise<CheckResult> {
+    const isGit = await this.gitOps.isInitialized();
+
+    return {
+      passed: isGit,
+      name: 'Git repository',
+      message: isGit ? 'Valid repository' : 'Not a git repository',
+      level: isGit ? 'info' : 'error',
+      suggestions: !isGit ? ['Run "git init" first'] : undefined,
+    };
+  }
+
+  private async checkHansoloInitialized(): Promise<CheckResult> {
+    try {
+      await this.sessionRepo.listSessions();
+      return {
+        passed: true,
+        name: 'Han-solo initialized',
+        level: 'info',
+      };
+    } catch {
+      return {
+        passed: false,
+        name: 'Han-solo initialized',
+        message: 'Not initialized',
+        level: 'error',
+        suggestions: ['Run "hansolo init" first'],
+      };
+    }
+  }
+}
+
+/**
+ * Post-flight checks for cleanup command
+ */
+class CleanupPostFlightChecks extends PostFlightChecks {
+  constructor(
+    private gitOps: GitOperations,
+    private sessionRepo: SessionRepository,
+    private branchesDeleted: string[]
+  ) {
+    super();
+    this.setupVerifications();
+  }
+
+  private setupVerifications(): void {
+    this.addVerification(async () => this.verifyOnMainBranch());
+    this.addVerification(async () => this.verifyOrphanedBranchesRemoved());
+    this.addVerification(async () => this.verifyNoStaleSessions());
+    this.addVerification(async () => this.verifyWorkingDirectoryClean());
+  }
+
+  private async verifyOnMainBranch(): Promise<CheckResult> {
+    const currentBranch = await this.gitOps.getCurrentBranch();
+    const isMain = currentBranch === 'main' || currentBranch === 'master';
+
+    return {
+      passed: isMain,
+      name: 'On main branch',
+      message: currentBranch,
+      level: isMain ? 'info' : 'warning',
+    };
+  }
+
+  private async verifyOrphanedBranchesRemoved(): Promise<CheckResult> {
+    const branches = await this.gitOps.listBranches();
+    const stillExists = this.branchesDeleted.filter(b => branches.includes(b));
+
+    return {
+      passed: stillExists.length === 0,
+      name: 'Orphaned branches removed',
+      message: stillExists.length > 0
+        ? `${stillExists.length} branches still exist`
+        : `${this.branchesDeleted.length} branches deleted`,
+      level: stillExists.length > 0 ? 'warning' : 'info',
+    };
+  }
+
+  private async verifyNoStaleSessions(): Promise<CheckResult> {
+    const sessions = await this.sessionRepo.listSessions({ all: true });
+    const stale = sessions.filter(s =>
+      s.currentState === 'COMPLETE' || s.currentState === 'ABORTED'
+    );
+
+    return {
+      passed: stale.length === 0,
+      name: 'No stale sessions',
+      message: stale.length > 0 ? `${stale.length} stale sessions remain` : undefined,
+      level: stale.length > 0 ? 'warning' : 'info',
+    };
+  }
+
+  private async verifyWorkingDirectoryClean(): Promise<CheckResult> {
+    const isClean = await this.gitOps.isClean();
+
+    return {
+      passed: isClean,
+      name: 'Working directory clean',
+      level: isClean ? 'info' : 'warning',
+      message: isClean ? undefined : 'Uncommitted changes present',
+    };
+  }
+}
+
+export class CleanupCommand implements CommandHandler {
   name = 'hansolo:cleanup';
-  description = 'Clean up completed sessions and branches';
+  description = 'Clean up completed sessions and orphaned branches';
 
   private sessionRepo: SessionRepository;
   private gitOps: GitOperations;
   private console: ConsoleOutput;
-  private progress: ProgressIndicator;
-  private box: BoxFormatter;
+  private progress: WorkflowProgress;
+  private branchesDeleted: string[] = [];
 
   constructor() {
     this.sessionRepo = new SessionRepository();
     this.gitOps = new GitOperations();
     this.console = new ConsoleOutput();
-    this.progress = new ProgressIndicator();
-    this.box = new BoxFormatter();
+    this.progress = new WorkflowProgress();
   }
 
   async execute(args: string[]): Promise<void> {
     try {
-      // Show banner
-      this.console.printBanner('🧹 han-solo Cleanup');
-
-      // Parse options
       const options = this.parseOptions(args);
 
-      // Sync main branch first (critical for post-PR cleanup)
+      // Run pre-flight checks
+      const preFlightChecks = new CleanupPreFlightChecks(
+        this.gitOps,
+        this.sessionRepo
+      );
+
+      const preFlightPassed = await preFlightChecks.runChecks({
+        command: 'cleanup',
+        options,
+      });
+
+      if (!preFlightPassed) {
+        this.console.errorMessage('\n❌ Pre-flight checks failed - aborting cleanup');
+        return;
+      }
+
+      // Display ASCII art banner
+      this.console.info(getBanner('cleanup'));
+
+      // Sync main branch first
       if (!options.noSync) {
         await this.syncMainBranch(options);
       }
 
-      // Start cleanup process
-      this.progress.start('Analyzing cleanup targets...');
-
       // Find cleanup candidates
+      const steps = [
+        {
+          name: 'Analyzing cleanup targets',
+          action: async () => {},
+        },
+      ];
+
+      await this.progress.runSteps(steps);
+
       const candidates = await this.findCleanupCandidates(options);
 
       if (candidates.sessions.length === 0 && candidates.branches.length === 0) {
-        this.progress.succeed('No cleanup needed - everything is tidy!');
+        this.console.successMessage('✨ No cleanup needed - everything is tidy!');
+
+        // Run post-flight even if nothing to clean
+        const postFlightChecks = new CleanupPostFlightChecks(
+          this.gitOps,
+          this.sessionRepo,
+          []
+        );
+        await postFlightChecks.runVerifications({
+          command: 'cleanup',
+          options,
+        });
+
         return;
       }
 
-      this.progress.stop();
-
       // Show cleanup summary
-      await this.showCleanupSummary(candidates);
+      this.showCleanupSummary(candidates);
 
       // Confirm cleanup
-      if (!options.force) {
+      if (!options.force && !options.yes) {
         const confirmed = await this.console.confirm(
-          'Proceed with cleanup? This action cannot be undone.'
+          '\nProceed with cleanup? This action cannot be undone.'
         );
 
         if (!confirmed) {
@@ -68,11 +220,27 @@ export class HansoloCleanupCommand implements CommandHandler {
       // Perform cleanup
       await this.performCleanup(candidates, options);
 
-      this.console.success('✅ Cleanup completed successfully');
+      // Run post-flight verifications
+      const postFlightChecks = new CleanupPostFlightChecks(
+        this.gitOps,
+        this.sessionRepo,
+        this.branchesDeleted
+      );
+
+      this.console.info('');
+      await postFlightChecks.runVerifications({
+        command: 'cleanup',
+        options,
+      });
+
+      this.console.info('');
+      this.console.successMessage('✅ Cleanup completed successfully');
+      this.console.infoMessage('🎉 Repository is clean and ready for the next feature!');
 
     } catch (error) {
-      this.progress.fail('Cleanup failed');
-      this.console.error('Failed to perform cleanup', error as Error);
+      this.console.errorMessage(
+        `Cleanup failed: ${error instanceof Error ? error.message : 'Unknown error'}`
+      );
       throw error;
     }
   }
@@ -84,6 +252,7 @@ export class HansoloCleanupCommand implements CommandHandler {
       orphaned: true,
       dryRun: false,
       force: false,
+      yes: false,
       days: 30,
       noSync: false,
     };
@@ -98,6 +267,10 @@ export class HansoloCleanupCommand implements CommandHandler {
       case '--force':
       case '-f':
         options.force = true;
+        break;
+      case '--yes':
+      case '-y':
+        options.yes = true;
         break;
       case '--all':
         options.all = true;
@@ -125,15 +298,10 @@ export class HansoloCleanupCommand implements CommandHandler {
     return options;
   }
 
-  /**
-   * Sync main branch with remote before cleanup
-   * Critical for ensuring local main has squashed PR commits
-   */
   private async syncMainBranch(options: CleanupOptions): Promise<void> {
     try {
-      this.console.info('📥 Syncing main branch with remote...');
+      this.console.subheader('📥 Syncing Main Branch');
 
-      // Get current branch and status
       const currentBranch = await this.gitOps.getCurrentBranch();
       const status = await this.gitOps.getStatus();
       const hasUncommittedChanges = !status.isClean;
@@ -141,7 +309,6 @@ export class HansoloCleanupCommand implements CommandHandler {
       let didStash = false;
       let wasOnFeatureBranch = false;
 
-      // Stash uncommitted changes if any
       if (hasUncommittedChanges) {
         this.progress.start('Stashing uncommitted changes...');
         await this.gitOps.stash('han-solo cleanup: auto-stash');
@@ -149,7 +316,6 @@ export class HansoloCleanupCommand implements CommandHandler {
         this.progress.succeed('Changes stashed');
       }
 
-      // Switch to main if not already there
       if (currentBranch !== 'main' && currentBranch !== 'master') {
         wasOnFeatureBranch = true;
         this.progress.start('Switching to main branch...');
@@ -157,50 +323,46 @@ export class HansoloCleanupCommand implements CommandHandler {
         this.progress.succeed('Switched to main');
       }
 
-      // Fetch latest from remote
       this.progress.start('Fetching latest from origin...');
       await this.gitOps.fetch('origin', 'main');
       this.progress.succeed('Fetched latest');
 
-      // Pull changes (fast-forward only to avoid conflicts)
       this.progress.start('Pulling latest changes...');
       try {
         await this.gitOps.pull('origin', 'main');
         this.progress.succeed('Main branch synced ✓');
-      } catch (error) {
+      } catch (_error) {
         this.progress.fail('Pull failed - may need manual rebase');
         this.console.warn('Unable to fast-forward main. You may need to manually sync.');
       }
 
-      // Return to original branch if we switched
       if (wasOnFeatureBranch && !options.branchesOnly) {
         this.progress.start(`Returning to ${currentBranch}...`);
         try {
           await this.gitOps.checkoutBranch(currentBranch);
           this.progress.succeed(`Back on ${currentBranch}`);
-        } catch (error) {
+        } catch (_error) {
           this.console.warn(`Could not return to ${currentBranch} - branch may have been merged`);
           this.console.info('Staying on main branch');
         }
       }
 
-      // Restore stashed changes
       if (didStash) {
         this.progress.start('Restoring stashed changes...');
         try {
           await this.gitOps.stashPop();
           this.progress.succeed('Changes restored');
-        } catch (error) {
+        } catch (_error) {
           this.console.warn('Could not auto-restore stashed changes');
           this.console.info('Run "git stash pop" manually to restore your changes');
         }
       }
 
-      this.console.success('✓ Main branch is now in sync with remote\n');
+      this.console.info('');
 
     } catch (error) {
       this.console.error('Failed to sync main branch', error as Error);
-      this.console.warn('Continuing with cleanup anyway...');
+      this.console.warn('Continuing with cleanup anyway...\n');
     }
   }
 
@@ -208,33 +370,29 @@ export class HansoloCleanupCommand implements CommandHandler {
     const candidates: CleanupCandidates = {
       sessions: [],
       branches: [],
-      locks: [],
     };
 
     if (!options.branchesOnly) {
-      // Find expired sessions
-      if (options.expired || options.all) {
-        const sessions = await this.sessionRepo.listSessions();
-        const now = new Date();
+      const sessions = await this.sessionRepo.listSessions({ all: true });
+      const now = new Date();
+      const threshold = new Date();
+      threshold.setDate(threshold.getDate() - options.days);
 
-        for (const session of sessions) {
+      for (const session of sessions) {
+        // Expired sessions
+        if (options.expired || options.all) {
           const expiresAt = new Date(session.expiresAt);
           if (expiresAt < now) {
             candidates.sessions.push({
               session,
               reason: 'Expired',
             });
+            continue;
           }
         }
-      }
 
-      // Find completed sessions older than threshold
-      if (options.completed || options.all) {
-        const sessions = await this.sessionRepo.listSessions();
-        const threshold = new Date();
-        threshold.setDate(threshold.getDate() - options.days);
-
-        for (const session of sessions) {
+        // Completed/aborted sessions
+        if (options.completed || options.all) {
           if (session.currentState === 'COMPLETE' || session.currentState === 'ABORTED') {
             const updatedAt = new Date(session.updatedAt);
             if (updatedAt < threshold) {
@@ -249,69 +407,58 @@ export class HansoloCleanupCommand implements CommandHandler {
     }
 
     if (!options.sessionsOnly) {
-      // Find orphaned branches (branches without sessions)
-      if (options.orphaned || options.all) {
-        const branches = await this.gitOps.listBranches();
-        const sessions = await this.sessionRepo.listSessions();
-        const sessionBranches = new Set(sessions.map(s => s.branchName));
+      const branches = await this.gitOps.listBranches();
+      const sessions = await this.sessionRepo.listSessions({ all: true });
+      const sessionBranches = new Set(sessions.map(s => s.branchName));
 
-        for (const branch of branches) {
-          if (!sessionBranches.has(branch) && branch !== 'main' && branch !== 'master') {
-            candidates.branches.push({
-              name: branch,
-              reason: 'No associated session',
-            });
-          }
+      for (const branch of branches) {
+        if (!sessionBranches.has(branch) && branch !== 'main' && branch !== 'master') {
+          candidates.branches.push({
+            name: branch,
+            reason: 'No associated session',
+          });
         }
       }
     }
 
-    // Find stale lock files
-    const locks = await this.findStaleLocks();
-    candidates.locks = locks;
-
     return candidates;
   }
 
-  private async showCleanupSummary(candidates: CleanupCandidates): Promise<void> {
-    const lines: string[] = [];
+  private showCleanupSummary(candidates: CleanupCandidates): void {
+    this.console.subheader('📋 Cleanup Summary');
 
     if (candidates.sessions.length > 0) {
-      lines.push(chalk.yellow(`Sessions to remove: ${candidates.sessions.length}`));
+      this.console.info(chalk.yellow(`\nSessions to remove: ${candidates.sessions.length}`));
       candidates.sessions.slice(0, 5).forEach(item => {
-        lines.push(`  • ${item.session.branchName} (${item.reason})`);
+        this.console.dim(`  • ${item.session.branchName} (${item.reason})`);
       });
       if (candidates.sessions.length > 5) {
-        lines.push(`  ... and ${candidates.sessions.length - 5} more`);
+        this.console.dim(`  ... and ${candidates.sessions.length - 5} more`);
       }
     }
 
     if (candidates.branches.length > 0) {
-      lines.push(chalk.yellow(`\nBranches to delete: ${candidates.branches.length}`));
+      this.console.info(chalk.yellow(`\nBranches to delete: ${candidates.branches.length}`));
       candidates.branches.slice(0, 5).forEach(item => {
-        lines.push(`  • ${item.name} (${item.reason})`);
+        this.console.dim(`  • ${item.name} (${item.reason})`);
       });
       if (candidates.branches.length > 5) {
-        lines.push(`  ... and ${candidates.branches.length - 5} more`);
+        this.console.dim(`  ... and ${candidates.branches.length - 5} more`);
       }
     }
-
-    if (candidates.locks.length > 0) {
-      lines.push(chalk.yellow(`\nLock files to remove: ${candidates.locks.length}`));
-    }
-
-    this.box.printBox('Cleanup Summary', lines.join('\n'));
   }
 
   private async performCleanup(candidates: CleanupCandidates, options: CleanupOptions): Promise<void> {
     if (options.dryRun) {
-      this.console.info('🔍 Dry run mode - no changes will be made');
+      this.console.info('\n🔍 Dry run mode - no changes will be made');
       return;
     }
 
+    this.console.info('');
+
     // Clean up sessions
     if (candidates.sessions.length > 0) {
-      this.progress.start('Removing expired sessions...');
+      this.progress.start('Removing stale sessions...');
       for (const item of candidates.sessions) {
         if (item.session?.id) {
           await this.sessionRepo.deleteSession(item.session.id);
@@ -326,90 +473,24 @@ export class HansoloCleanupCommand implements CommandHandler {
       for (const item of candidates.branches) {
         try {
           await this.gitOps.deleteBranch(item.name, options.force);
+          this.branchesDeleted.push(item.name);
         } catch (error) {
           this.console.warn(`Failed to delete branch ${item.name}: ${error}`);
         }
       }
-      this.progress.succeed(`Deleted ${candidates.branches.length} branches`);
+      this.progress.succeed(`Deleted ${this.branchesDeleted.length} branches`);
     }
 
-    // Clean up lock files
-    if (candidates.locks.length > 0) {
-      this.progress.start('Removing stale locks...');
-      for (const lock of candidates.locks) {
-        await this.removeLockFile(lock);
-      }
-      this.progress.succeed(`Removed ${candidates.locks.length} lock files`);
-    }
-
-    // Clean up audit logs if requested
-    if (options.all) {
-      await this.cleanupAuditLogs(options.days);
-    }
-  }
-
-  private async findStaleLocks(): Promise<string[]> {
-    const fs = await import('fs/promises');
-    const path = await import('path');
-
-    try {
-      const lockDir = '.hansolo/locks';
-      const files = await fs.readdir(lockDir);
-      const locks: string[] = [];
-
-      for (const file of files) {
-        if (file.endsWith('.lock')) {
-          const lockPath = path.join(lockDir, file);
-          const stat = await fs.stat(lockPath);
-          const ageInHours = (Date.now() - stat.mtime.getTime()) / (1000 * 60 * 60);
-
-          // Consider locks older than 24 hours as stale
-          if (ageInHours > 24) {
-            locks.push(lockPath);
-          }
-        }
-      }
-
-      return locks;
-    } catch {
-      return [];
-    }
-  }
-
-  private async removeLockFile(lockPath: string): Promise<void> {
-    const fs = await import('fs/promises');
-    try {
-      await fs.unlink(lockPath);
-    } catch {
-      // Ignore errors if file doesn't exist
-    }
-  }
-
-  private async cleanupAuditLogs(daysToKeep: number): Promise<void> {
-    const fs = await import('fs/promises');
-    const path = await import('path');
-
-    try {
-      const auditDir = '.hansolo/audit';
-      const threshold = new Date();
-      threshold.setDate(threshold.getDate() - daysToKeep);
-
-      const months = await fs.readdir(auditDir);
-      for (const month of months) {
-        const monthDir = path.join(auditDir, month);
-        const stat = await fs.stat(monthDir);
-
-        if (stat.isDirectory() && stat.mtime < threshold) {
-          await fs.rm(monthDir, { recursive: true });
-        }
-      }
-    } catch {
-      // Ignore errors if audit directory doesn't exist
+    // Switch to main if we deleted current branch
+    const currentBranch = await this.gitOps.getCurrentBranch();
+    if (this.branchesDeleted.includes(currentBranch)) {
+      this.progress.start('Switching to main branch...');
+      await this.gitOps.checkoutBranch('main');
+      this.progress.succeed('Switched to main');
     }
   }
 
   validate(_args: string[]): boolean {
-    // All arguments are optional
     return true;
   }
 }
@@ -420,6 +501,7 @@ interface CleanupOptions {
   orphaned: boolean;
   dryRun: boolean;
   force: boolean;
+  yes: boolean;
   days: number;
   noSync: boolean;
   all?: boolean;
@@ -436,5 +518,4 @@ interface CleanupCandidates {
     name: string;
     reason: string;
   }>;
-  locks: string[];
 }
