--- src/commands/hansolo-ship.d.ts
@@ -1,30 +1,0 @@
-export declare class ShipCommand {
-    private output;
-    private progress;
-    private sessionRepo;
-    private gitOps;
-    private configManager;
-    private stateMachine;
-    private githubIntegration;
-    constructor(basePath?: string);
-    execute(options?: {
-        message?: string;
-        push?: boolean;
-        createPR?: boolean;
-        merge?: boolean;
-        force?: boolean;
-        yes?: boolean;
-    }): Promise<void>;
-    private determineNextAction;
-    private performCommit;
-    private performPush;
-    private performCreatePR;
-    private performMerge;
-    private performComplete;
-    private promptForCommitMessage;
-    private confirmAction;
-    private generatePRDescription;
-    private showCompletionSummary;
-    quickShip(): Promise<void>;
-}
-//# sourceMappingURL=hansolo-ship.d.ts.map
